fix(manager-components): refacto tiles input/stepper/tabs
